Add regenerate character option

+++ app/generator/characters.py
@@ -1,4 +1,4 @@
-from flask import render_template, redirect, url_for, jsonify
+from flask import render_template, jsonify, abort
 from app.generator.models import Character, Name, Skill, Attribute, Item
 from app.account.models import User
 from app.shared.models import db
@@ -17,9 +17,10 @@ def remove_character(character_id):
 
     # TODO: restrict other users access
     ch = db.one_or_404(db.select(Character)
-                        .filter_by(id=character_id)
-                        .join(User)
-                        .filter_by(id=current_user.id))
+                        .filter_by(id=character_id))
+
+    if (not current_user.id == ch.user.id and not current_user.moderator):
+        abort(404)  
 
     db.session.delete(ch)
     db.session.commit()
@@ -29,9 +30,11 @@ def remove_character(character_id):
 def show_character(character_id):
 
     ch = db.one_or_404(db.select(Character)
-                        .filter_by(id=character_id)
-                        .join(User)
-                        .filter_by(id=current_user.id))
+                        .filter_by(id=character_id))
+
+    if (not current_user.id == ch.user.id and not current_user.moderator):
+        abort(404)  
+
     character_data = {}  
     character_data['id'] = ch.id 
     character_data['timestamp'] = ch.timestamp.strftime("%d.%m.%Y %H:%M:%S")
@@ -56,7 +59,6 @@ def show_character(character_id):
 
     print("character_data")
     print(character_data)
-    #return jsonify(character_data)
     return render_template("character.html", character=character_data, title="RPG Generator",
                            user=current_user)
 
@@ -72,7 +74,6 @@ def generate_character():
     attributes_list = [a for a in attributes]
     items_list = [i for i in items]
 
-    if not name: return
     print("name")
     print(name.name)
 
@@ -87,20 +88,35 @@ def generate_character():
     db.session.add(c)
     db.session.commit()
 
-    result = []  
-    for s in skills:  
-        skill_data = {}  
-        skill_data['id'] = s.id 
-        skill_data['name'] = s.name
-
-        result.append(skill_data)  
-
-    print("Generated skills:")
-    print(result)
-
-    #return make_response(jsonify({'should': "generate"}))
     return jsonify({'id': c.id})
 
 def regenerate_character(character_id):
-    print("character id:", character_id)
-    return jsonify({'id': 1})
+    c = db.one_or_404(db.select(Character)
+                        .filter_by(id=character_id))
+
+    if (not current_user.id == c.user.id and not current_user.moderator):
+        abort(404)  
+
+    name = db.first_or_404(db.select(Name).order_by(func.random()))
+
+    skills = db.session.execute(db.select(Skill).order_by(func.random()).limit(3)).scalars()
+    attributes = db.session.execute(db.select(Attribute).order_by(func.random()).limit(3)).scalars()
+    items = db.session.execute(db.select(Item).order_by(func.random()).limit(3)).scalars()
+
+    skills_list = [s for s in skills]
+    attributes_list = [a for a in attributes]
+    items_list = [i for i in items]
+
+    print("name")
+    print(name.name)
+
+    print(current_user.name)
+
+    c.name = name
+    c.skills = skills_list
+    c.attributes = attributes_list
+    c.items = items_list
+
+    db.session.commit()
+
+    return jsonify({'id': c.id})
